Save snapshot of original menu arguments and use it in transient name (#2)
If `menu` argument is not set for a menu, `wp_nav_menu()` sets one. This means that there is a difference between original arguments (that we use when checking for cache, on `pre_wp_nav_menu`) and those we use when we want to set cache (on `wp_nav_menu`). Because of this, plugin doesn't work for many cases and it can create opposite effect.

In #1, another example was shown where menu arguments can be modified in a similar way.

This change adds a hash of menu arguments as a property of that object and uses it when generating menu cache key.